Fixed issue 6 (https://code.google.com/p/midi-dot-net/issues/detail?id=6): Midi-dot-net doesn't work on Windows 8

+++ Midi/Win32API.cs
@@ -173,7 +173,7 @@ namespace Midi
         [StructLayout(LayoutKind.Sequential)]
         public struct HMIDIOUT
         {
-            public Int32 handle;
+            public IntPtr handle;
         }
 
         /// <summary>
@@ -182,7 +182,7 @@ namespace Midi
         [StructLayout(LayoutKind.Sequential)]
         public struct HMIDIIN
         {
-            public Int32 handle;
+            public IntPtr handle;
         }
 
         #endregion
